+++ Core/src/Domain/Mesh.cpp
@@ -50,7 +50,7 @@ namespace fstim
         return -1;
     }
 
-    Mesh::Mesh(int nCellsIn, int nFacesIn,  std::unique_ptr<Cell[]> cellsIn, std::unique_ptr<Face[]> facesIn, vecp::Vec2f lengthIn) 
+    Mesh::Mesh(int nCellsIn, int nFacesIn,  std::unique_ptr<Cell[]> cellsIn, std::unique_ptr<Face[]> facesIn, vecp::Vec2d lengthIn) 
         : nCells(nCellsIn), nFaces(nFacesIn), cells(std::move(cellsIn)), faces(std::move(facesIn)), length(lengthIn)
     {
         m_faceSets.resize(0);
